Add raw split/raw checksum/raw write sst (#10524)
* Add raw split/raw checksum/raw write sst


* Fix build


* fix warning


* fix clippy


* format


* add metric


* fix clippy


* fix metrics


* fix test


* remove todo


* update kvproto
