+++ components/engine_traits/Cargo.toml
@@ -23,6 +23,7 @@ prost-codec = [
   "txn_types/prost-codec",
   "file_system/prost-codec",
 ]
+failpoints = ["fail/failpoints"]
 
 [dependencies]
 error_code = { path = "../error_code", default-features = false }
@@ -38,6 +39,7 @@ slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debu
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "d592f88e4dbba5eb439998463054f1a44fbf17b9" }
 kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
 raft = { version = "0.6.0-alpha", default-features = false }
+fail = "0.4"
 
 [dev-dependencies]
 toml = "0.5"
