+++ components/engine_traits/src/lib.rs
@@ -259,6 +259,8 @@ extern crate tikv_alloc;
 #[macro_use]
 extern crate serde_derive;
 extern crate slog_global;
+#[macro_use(fail_point)]
+extern crate fail;
 
 // These modules contain traits that need to be implemented by engines, either
 // they are required by KvEngine or are an associated type of KvEngine. It is
