+++ components/sst_importer/src/errors.rs
@@ -107,6 +107,9 @@ pub enum Error {
 
     #[error("ingest file conflict")]
     FileConflict,
+
+    #[error("ttl is not enabled")]
+    TTLNotEnabled,
 }
 
 impl From<String> for Error {
@@ -148,6 +151,7 @@ impl ErrorCodeExt for Error {
             Error::Encryption(e) => e.error_code(),
             Error::CodecError(e) => e.error_code(),
             Error::FileConflict => error_code::sst_importer::FILE_CONFLICT,
+            Error::TTLNotEnabled => error_code::sst_importer::TTL_NOT_ENABLED,
         }
     }
 }
