+++ components/test_raftstore/src/node.rs
@@ -250,7 +250,7 @@ impl Simulator for NodeCluster {
 
         let importer = {
             let dir = Path::new(engines.kv.path()).join("import-sst");
-            Arc::new(SSTImporter::new(&cfg.import, dir, None).unwrap())
+            Arc::new(SSTImporter::new(&cfg.import, dir, None, false).unwrap())
         };
 
         let local_reader = LocalReader::new(engines.kv.clone(), store_meta.clone(), router.clone());
