+++ components/test_raftstore/src/server.rs
@@ -320,7 +320,7 @@ impl Simulator for ServerCluster {
         // Create import service.
         let importer = {
             let dir = Path::new(engines.kv.path()).join("import-sst");
-            Arc::new(SSTImporter::new(&cfg.import, dir, key_manager.clone()).unwrap())
+            Arc::new(SSTImporter::new(&cfg.import, dir, key_manager.clone(), false).unwrap())
         };
         let import_service = ImportSSTService::new(
             cfg.import.clone(),
