+++ src/import/mod.rs
@@ -18,7 +18,7 @@ mod sst_service;
 pub use self::sst_service::ImportSSTService;
 pub use sst_importer::Config;
 pub use sst_importer::{Error, Result};
-pub use sst_importer::{SSTImporter, SSTWriter};
+pub use sst_importer::{SSTImporter, TxnSSTWriter};
 
 use grpcio::{RpcStatus, RpcStatusCode};
 use std::fmt::Debug;
