+++ src/import/sst_service.rs
@@ -21,6 +21,8 @@ use kvproto::{errorpb, kvrpcpb::Context};
 #[cfg(feature = "prost-codec")]
 use kvproto::import_sstpb::write_request::*;
 #[cfg(feature = "protobuf-codec")]
+use kvproto::import_sstpb::RawWriteRequest_oneof_chunk as RawChunk;
+#[cfg(feature = "protobuf-codec")]
 use kvproto::import_sstpb::WriteRequest_oneof_chunk as Chunk;
 use kvproto::import_sstpb::*;
 
@@ -100,6 +102,7 @@ where
         let p = sst_meta_to_path(meta)?;
         Ok(slots.insert(p))
     }
+
     fn release_lock(task_slots: &Arc<Mutex<HashSet<PathBuf>>>, meta: &SstMeta) -> Result<bool> {
         let mut slots = task_slots.lock().unwrap();
         let p = sst_meta_to_path(meta)?;
@@ -137,6 +140,7 @@ where
             term: header.get_current_term(),
         })
     }
+
     fn check_write_stall(&self) -> Option<errorpb::Error> {
         if self.importer.get_mode() == SwitchMode::Normal
             && self
@@ -222,6 +226,68 @@ where
             Ok(resp)
         }
     }
+}
+
+#[macro_export]
+macro_rules! impl_write {
+    ($fn:ident, $req_ty:ident, $resp_ty:ident, $chunk_ty:ident, $writer_fn:ident) => {
+        fn $fn(
+            &mut self,
+            _ctx: RpcContext<'_>,
+            stream: RequestStream<$req_ty>,
+            sink: ClientStreamingSink<$resp_ty>,
+        ) {
+            let import = self.importer.clone();
+            let engine = self.engine.clone();
+            let (rx, buf_driver) =
+                create_stream_with_buffer(stream, self.cfg.stream_channel_window);
+            let mut rx = rx.map_err(Error::from);
+
+            let timer = Instant::now_coarse();
+            let label = stringify!($fn);
+            let handle_task = async move {
+                let res = async move {
+                    let first_req = rx.try_next().await?;
+                    let meta = match first_req {
+                        Some(r) => match r.chunk {
+                            Some($chunk_ty::Meta(m)) => m,
+                            _ => return Err(Error::InvalidChunk),
+                        },
+                        _ => return Err(Error::InvalidChunk),
+                    };
+
+                    let writer = match import.$writer_fn(&engine, meta) {
+                        Ok(w) => w,
+                        Err(e) => {
+                            error!("build writer failed {:?}", e);
+                            return Err(Error::InvalidChunk);
+                        }
+                    };
+                    let writer = rx
+                        .try_fold(writer, |mut writer, req| async move {
+                            let batch = match req.chunk {
+                                Some($chunk_ty::Batch(b)) => b,
+                                _ => return Err(Error::InvalidChunk),
+                            };
+                            writer.write(batch)?;
+                            Ok(writer)
+                        })
+                        .await?;
+
+                    writer.finish().map(|metas| {
+                        let mut resp = $resp_ty::default();
+                        resp.set_metas(metas.into());
+                        resp
+                    })
+                }
+                .await;
+                crate::send_rpc_response!(res, sink, label, timer);
+            };
+
+            self.threads.spawn_ok(buf_driver);
+            self.threads.spawn_ok(handle_task);
+        }
+    };
 }
 
 impl<E, Router> ImportSst for ImportSSTService<E, Router>
@@ -522,64 +588,6 @@ where
         ctx.spawn(ctx_task);
     }
 
-    fn write(
-        &mut self,
-        _ctx: RpcContext<'_>,
-        stream: RequestStream<WriteRequest>,
-        sink: ClientStreamingSink<WriteResponse>,
-    ) {
-        let label = "write";
-        let timer = Instant::now_coarse();
-        let import = self.importer.clone();
-        let engine = self.engine.clone();
-        let (rx, buf_driver) = create_stream_with_buffer(stream, self.cfg.stream_channel_window);
-        let mut rx = rx.map_err(Error::from);
-
-        let handle_task = async move {
-            let res = async move {
-                let first_req = rx.try_next().await?;
-                let meta = match first_req {
-                    Some(r) => match r.chunk {
-                        Some(Chunk::Meta(m)) => m,
-                        _ => return Err(Error::InvalidChunk),
-                    },
-                    _ => return Err(Error::InvalidChunk),
-                };
-
-                let writer = match import.new_writer::<E>(&engine, meta) {
-                    Ok(w) => w,
-                    Err(e) => {
-                        error!("build writer failed {:?}", e);
-                        return Err(Error::InvalidChunk);
-                    }
-                };
-                let writer = rx
-                    .try_fold(writer, |mut writer, req| async move {
-                        let start = Instant::now_coarse();
-                        let batch = match req.chunk {
-                            Some(Chunk::Batch(b)) => b,
-                            _ => return Err(Error::InvalidChunk),
-                        };
-                        writer.write(batch)?;
-                        IMPORT_WRITE_CHUNK_DURATION.observe(start.saturating_elapsed_secs());
-                        Ok(writer)
-                    })
-                    .await?;
-
-                writer.finish().map(|metas| {
-                    let mut resp = WriteResponse::default();
-                    resp.set_metas(metas.into());
-                    resp
-                })
-            }
-            .await;
-            crate::send_rpc_response!(res, sink, label, timer);
-        };
-
-        self.threads.spawn_ok(buf_driver);
-        self.threads.spawn_ok(handle_task);
-    }
-
     fn duplicate_detect(
         &mut self,
         _ctx: RpcContext<'_>,
@@ -646,6 +654,16 @@ where
         };
         self.threads.spawn_ok(handle_task);
     }
+
+    impl_write!(write, WriteRequest, WriteResponse, Chunk, new_txn_writer);
+
+    impl_write!(
+        raw_write,
+        RawWriteRequest,
+        RawWriteResponse,
+        RawChunk,
+        new_raw_writer
+    );
 }
 
 // add error statistics from pb error response
