+++ src/server/metrics.rs
@@ -42,6 +42,7 @@ make_auto_flush_static_metric! {
         raw_batch_delete,
         raw_get_key_ttl,
         raw_compare_and_swap,
+        raw_checksum,
         unsafe_destroy_range,
         physical_scan_lock,
         register_lock_observer,
