+++ src/server/ttl/mod.rs
@@ -3,5 +3,5 @@
 mod ttl_checker;
 mod ttl_compaction_filter;
 
-pub use ttl_checker::{TTLChecker, Task as TTLCheckerTask};
+pub use ttl_checker::{check_ttl_and_compact_files, TTLChecker, Task as TTLCheckerTask};
 pub use ttl_compaction_filter::TTLCompactionFilterFactory;
