+++ src/server/ttl/ttl_compaction_filter.rs
@@ -3,13 +3,12 @@
 use std::ffi::CString;
 
 use crate::server::metrics::TTL_CHECKER_ACTIONS_COUNTER_VEC;
-use crate::storage::raw::ttl::current_ts;
 use engine_rocks::raw::{
     new_compaction_filter_raw, CompactionFilter, CompactionFilterContext, CompactionFilterDecision,
     CompactionFilterFactory, CompactionFilterValueType, DBCompactionFilter,
 };
 use engine_rocks::{RocksTtlProperties, RocksUserCollectedPropertiesNoRc};
-use engine_traits::util::get_expire_ts;
+use engine_traits::util::{get_expire_ts, ttl_current_ts};
 
 pub struct TTLCompactionFilterFactory;
 
@@ -18,7 +17,7 @@ impl CompactionFilterFactory for TTLCompactionFilterFactory {
         &self,
         context: &CompactionFilterContext,
     ) -> *mut DBCompactionFilter {
-        let current = current_ts();
+        let current = ttl_current_ts();
 
         let mut min_expire_ts = u64::MAX;
         for i in 0..context.file_numbers().len() {
@@ -79,60 +78,5 @@ impl CompactionFilter for TTLCompactionFilter {
         } else {
             CompactionFilterDecision::Keep
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use engine_traits::util::append_expire_ts;
-
-    use crate::config::DbConfig;
-    use crate::storage::kv::TestEngineBuilder;
-    use crate::storage::raw::ttl::TEST_CURRENT_TS;
-    use engine_rocks::raw::CompactOptions;
-    use engine_rocks::util::get_cf_handle;
-    use engine_traits::{MiscExt, Peekable, SyncMutable, CF_DEFAULT};
-
-    #[test]
-    fn test_ttl_compaction_filter() {
-        let mut cfg = DbConfig::default();
-        cfg.writecf.disable_auto_compactions = true;
-        let dir = tempfile::TempDir::new().unwrap();
-        let builder = TestEngineBuilder::new().path(dir.path()).ttl(true);
-        let engine = builder.build_with_cfg(&cfg).unwrap();
-        let kvdb = engine.get_rocksdb();
-
-        let key1 = b"zkey1";
-        let mut value1 = vec![0; 10];
-        append_expire_ts(&mut value1, 10);
-        kvdb.put_cf(CF_DEFAULT, key1, &value1).unwrap();
-        kvdb.flush_cf(CF_DEFAULT, true).unwrap();
-
-        let db = kvdb.as_inner();
-        let handle = get_cf_handle(db, CF_DEFAULT).unwrap();
-        db.compact_range_cf_opt(handle, &CompactOptions::new(), None, None);
-
-        assert!(kvdb.get_value_cf(CF_DEFAULT, key1).unwrap().is_none());
-
-        let key2 = b"zkey2";
-        let mut value2 = vec![0; 10];
-        append_expire_ts(&mut value2, TEST_CURRENT_TS + 20);
-        kvdb.put_cf(CF_DEFAULT, key2, &value2).unwrap();
-        let key3 = b"zkey3";
-        let mut value3 = vec![0; 10];
-        append_expire_ts(&mut value3, 20);
-        kvdb.put_cf(CF_DEFAULT, key3, &value3).unwrap();
-        kvdb.flush_cf(CF_DEFAULT, true).unwrap();
-
-        let key4 = b"zkey4";
-        let mut value4 = vec![0; 10];
-        append_expire_ts(&mut value4, 0);
-        kvdb.put_cf(CF_DEFAULT, key4, &value4).unwrap();
-        kvdb.flush_cf(CF_DEFAULT, true).unwrap();
-
-        db.compact_range_cf_opt(handle, &CompactOptions::new(), None, None);
-        assert!(kvdb.get_value_cf(CF_DEFAULT, key2).unwrap().is_some());
-        assert!(kvdb.get_value_cf(CF_DEFAULT, key3).unwrap().is_none());
-        assert!(kvdb.get_value_cf(CF_DEFAULT, key4).unwrap().is_some());
     }
 }
