+++ src/storage/metrics.rs
@@ -141,6 +141,7 @@ make_auto_flush_static_metric! {
         raw_get_key_ttl,
         raw_compare_and_swap,
         raw_atomic_store,
+        raw_checksum,
     }
 
     pub label_enum CommandStageKind {
