+++ tests/failpoints/cases/mod.rs
@@ -25,3 +25,4 @@ mod test_stale_read;
 mod test_storage;
 mod test_transaction;
 mod test_transfer_leader;
+mod test_ttl;
